feat: make some changes on Keyring and Wallet
