+++ src/PlugKeyRing/plugKeyRing.test.ts
@@ -15,14 +15,14 @@ import {
 
 import PlugKeyRing from '.';
 import { ERRORS } from '../errors';
-import PlugWallet from '../PlugWallet';
+import PlugWallet from '../PlugWallet/base';
 import { createAgent } from '../utils/dfx';
 import store from '../utils/storage/mock';
 import { getAccountId } from '../utils/account';
 import { DEFAULT_MAINNET_ASSETS, TOKENS } from '../constants/tokens';
 import { GetTransactionsResponse } from '../interfaces/transactions';
 import { Mainnet } from './modules/NetworkModule/Network';
-import { Types } from '../utils/account/constants';
+import { IDENTITY_TYPES } from '../utils/account/constants';
 import { createAccountFromMnemonic } from '../utils/account';
 
 const mockSendICP = jest.fn();
@@ -143,14 +143,14 @@ const createManyTransactions = (): GetTransactionsResponse => {
 };
 
 describe('Plug KeyRing', () => {
-  const { identity } = createAccountFromMnemonic(TEST_MNEMONIC,0);
+  const { identity } = createAccountFromMnemonic(TEST_MNEMONIC, 0);
   const testWallet = new PlugWallet({
     name: 'test',
     walletId: '0',
     orderNumber: 0,
     fetch,
     network: new Mainnet({}, fetch),
-    type: Types.mnemonic,
+    type: IDENTITY_TYPES.mnemonic,
     identity,
   });
   let keyRing: PlugKeyRing;
@@ -658,9 +658,9 @@ describe('Plug KeyRing', () => {
     });
 
     test('get error with invalid wallet numbers', async () => {
-      await expect(keyRing.getTransactions({ subaccount: '-2' })).rejects.toThrow(
-        ERRORS.INVALID_WALLET_NUMBER
-      );
+      await expect(
+        keyRing.getTransactions({ subaccount: '-2' })
+      ).rejects.toThrow(ERRORS.INVALID_WALLET_NUMBER);
       await expect(
         keyRing.getTransactions({ subaccount: 'walletsCreated + 2' })
       ).rejects.toThrow(ERRORS.INVALID_WALLET_NUMBER);
